Update GetGameList.kt
Make type consistent with the rest of endpoints.

+++ src/main/kotlin/org/retroachivements/api/data/pojo/system/GetGameList.kt
@@ -22,7 +22,7 @@ class GetGameList {
             val imageIcon: String,
 
             @SerializedName("NumAchievements")
-            val numAchievements: Long,
+            val numAchievements: Int,
 
             @SerializedName("NumLeaderboards")
             val numLeaderboards: Long,
